+++ src/pages/Authorization/SignUp/components/SignUpForm.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { getLocalStorageItem, setLocalStorageItem } from '../../../../utils/localStorage';
-import { Input } from '../../components/inputComponent';
+import { Input } from '../../../../components/Input/Input';
 
 const defaultUser = {
   name: '',
